finance_doc_analyzer/ai_service/requirements.txt
# Core dependencies - fixed versions only where necessary
crewai==0.130.0
fastapi>=0.100.0
uvicorn>=0.20.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
aiofiles>=23.1.0

# LangChain and OpenAI dependencies
langchain-openai>=0.1.0
langchain-community>=0.2.0
openai>=1.30.0

# PDF processing
pypdf2>=3.0.0

# Optional dependencies for queue worker (Redis/Celery)
redis>=5.0.0
celery>=5.3.0
flower>=2.0.0

# Optional dependencies for database integration  
sqlalchemy>=2.0.0
alembic>=1.13.0

# CrewAI Tools
crewai-tools>=0.40.0